Updated params usage
-Added vendor packages
-Fixed tests
-Added READMEs

+++ pt-mongodb-summary/runtests.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-export BASEDIR=$(git rev-parse --show-toplevel)
+export BASEDIR=$(git rev-parse --show-toplevel)/pt-mongodb-summary
 if [ ! -d "${GOPATH}/src/github.com/golang/mock/gomock" ] 
 then
     echo "Installing gomock & mockgen"
@@ -10,8 +10,10 @@ cd $BASEDIR
 
 withmock go test -v -coverprofile=${BASEDIR}/coverage.out
 
-#TODO withmock is producing a weird behavior in os.Getwd and coverage. 
-sed -i 's/@ithub.com/github.com/' coverage.out
-go tool cover -func=coverage.out
 
-rm coverage.out
+if [ -f coverage.out ]
+then
+    sed -i 's/@ithub.com/github.com/' coverage.out
+    go tool cover -func=coverage.out
+    rm coverage.out
+fi
